init git lfs to track sample data for now
