+++ .gitattributes
@@ -1,0 +1,1 @@
+*.avi filter=lfs diff=lfs merge=lfs -text
