+++ .gitignore
@@ -23,3 +23,6 @@ share/python-wheels/
 .installed.cfg
 *.egg
 MANIFEST
+
+animal_soup/.pytest_cache/
+tests/tmp
