+++ tests/sample_data/M232Slc17a7_Gtacr2/20170306/M232_20170306_v017.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e19425b6095718bdbcdc4fb79bd8d48efe5ceddfc7f2ae0228e5d73a17cad997
+size 25980126
