+++ tests/sample_data/M232Slc17a7_Gtacr2/20170306/M232_20170306_v024.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:495cde15d8422c0e50c31390bb5af1ed7a117eaaa1e288b951e2f96c3b2c7410
+size 33589226
