+++ tests/sample_data/M232Slc17a7_Gtacr2/20170306/M232_20170306_v025.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8841fd5acde94c20fdec08b64070555e3fc2fc59bd72a28eebca3da96dede0ca
+size 33690626
