+++ tests/sample_data/M232Slc17a7_Gtacr2/20170307/M232_20170307_v022.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:22a989802f220db38f6f892912ad508df6b08fa81432755bc6d15438439283db
+size 32940144
