+++ tests/sample_data/M232Slc17a7_Gtacr2/20170307/M232_20170307_v029.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0089ee8aa88bd49ab458eea1e5d1cae93da1cb362e1816f647c5eec38fb56a2a
+size 33116390
