+++ tests/sample_data/M232Slc17a7_Gtacr2/20170307/M232_20170307_v047.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:3cc5fd164bb82b380163826fb9213955dd79a0c175a2a5076dde82583e1243bc
+size 26540364
