+++ tests/sample_data/M234Slc17a7_Gtacr2/20170328/M234_20170328_v002.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ea6c85fd993aafdec331413440047b16b0907e3940bb77d2ad959a089fba3c49
+size 19462442
