+++ tests/sample_data/M234Slc17a7_Gtacr2/20170328/M234_20170328_v004.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:60ecc9700d3dfae4cab1118b297ad34894b4b9fc53f0ee88129e22db70a23823
+size 19129182
