+++ tests/sample_data/M234Slc17a7_Gtacr2/20170328/M234_20170328_v005.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2df16e4e712bf25a8c5ef01fd8ad7cda019ccfd3a3fbd1da9489b1245cc35add
+size 29587026
