+++ tests/sample_data/M234Slc17a7_Gtacr2/20170329/M234_20170329_v001.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ade1a32379296565dcec04151251cb526c26e2355bc66a022ece26158dcf5516
+size 25420120
