+++ tests/sample_data/M234Slc17a7_Gtacr2/20170329/M234_20170329_v002.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d14189c97ea52b1c26b0807dc1d34dd93f482dcbdb94200fef73e2bbfdaf7ced
+size 25457174
