+++ tests/sample_data/M234Slc17a7_Gtacr2/20170329/M234_20170329_v003.avi
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d88d6897586c0fe43293f822fc677d4c8535cbe091e463138e74278f83c953ef
+size 25653166
